[DUT-875]: packages/api 초안 분리
