+++ packages/api/src/account/index.ts
@@ -1,0 +1,2 @@
+export * from './contracts';
+export * from './create-account-api';
