+++ packages/api/src/ward/index.ts
@@ -1,0 +1,2 @@
+export * from './contracts';
+export * from './create-ward-api';
