+++ packages/api/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../config/typescript/base.json",
+    "compilerOptions": {
+        "composite": true,
+        "noEmit": true
+    },
+    "include": ["src"]
+}
